memory management improvements, bug fixes
