+++ Data.cs
@@ -37,7 +37,7 @@ public class Data
         cannon.cost = 50;
         cannon.name = "Cannon";
         cannon.shotCount = 3;
-        cannon.controlTrajectory = true;
+        cannon.controlTrajectory = false;
         cannon.inPossesion = possesionArray[WEAPON_CANNON_ID];
         cannon.projectileTexture = ResourceLoader.Load("res://textures/CannonProjectile.png") as Texture;
         cannon.projectileWeight = 2;
@@ -48,7 +48,7 @@ public class Data
         GD.Print("weapons loaded");
     }
 
-    public static void AddToPossesion(int weaponIndex)
+    public static void AddWeaponToPossesion(int weaponIndex)
     {
         weapons[weaponIndex].inPossesion = true;
