+++ EnemyCounterPanel.tscn
@@ -1,28 +1,33 @@
-[gd_scene load_steps=4 format=2]
+[gd_scene load_steps=5 format=2]
 
-[ext_resource path="res://res/knight.svg" type="Texture" id=1]
+[ext_resource path="res://res/peasant_preview.svg" type="Texture" id=1]
 [ext_resource path="res://res/cross.png" type="Texture" id=2]
 [ext_resource path="res://EnemyCounterPanel.cs" type="Script" id=3]
+
+[sub_resource type="AtlasTexture" id=1]
+flags = 4
+atlas = ExtResource( 1 )
+region = Rect2( 1035.02, 216.419, 1710.27, 2153.59 )
 
 [node name="EnemyCounterPanel" type="Node2D"]
 script = ExtResource( 3 )
 
 [node name="EnemyScoreImage" type="TextureRect" parent="."]
-margin_left = 0.0736847
-margin_top = 0.29142
-margin_right = 32.0737
-margin_bottom = 58.2914
-texture = ExtResource( 1 )
+anchor_left = 0.5
+anchor_right = 0.526
+margin_left = -4.0
+margin_top = 3.0
+margin_right = 43.0
+margin_bottom = 59.0
+texture = SubResource( 1 )
 expand = true
 __meta__ = {
 "_edit_use_anchors_": false
 }
 
 [node name="Cross" type="TextureRect" parent="."]
-margin_left = 3.27278
-margin_top = -0.467545
-margin_right = 36.2728
-margin_bottom = 59.5325
+margin_right = 46.0
+margin_bottom = 60.0
 texture = ExtResource( 2 )
 expand = true
 __meta__ = {
