+++ PropertyKeys.cs
@@ -1,6 +1,8 @@
 public class PropertyKeys {
     public const string COIN_COUNT = "Coin_count";
     public const string WEAPON_POSSESION_LIST = "WeaponPossesion";
-    public const string PASSED_LEVELS_LIST = "PassedLevelsList";
+    public const string AVAILABLE_LEVELS = "AvailableLevel";
     public const string CURRENT_WEAPON = "CurrentWeapon";
+
+    public const string CURRENT_LEVEL = "currentLevel";
 }
