+++ PropertyLabel.cs
@@ -3,9 +3,16 @@ using System;
 
 public class PropertyLabel : Label
 {
+    private int propertyIndex;
+
+    ~PropertyLabel()
+    {
+        StorageManager.UnsubscribeToPropertyChange(PropertyKeys.COIN_COUNT, propertyIndex);
+    }
+
     public override void _Ready()
     {
-        StorageManager.SubscibeToPropertyChange(PropertyKeys.COIN_COUNT, (key, value) => Text = value.ToString());
+        propertyIndex = StorageManager.SubscibeToPropertyChange(PropertyKeys.COIN_COUNT, (key, value) => Text = value.ToString());
 
         Text = StorageManager.GetInt(PropertyKeys.COIN_COUNT).ToString();
     }
