+++ export_presets.cfg
@@ -22,8 +22,8 @@ application/provisioning_profile_uuid_debug=""
 application/code_sign_identity_debug="iPhone Developer"
 application/export_method_debug=1
 application/provisioning_profile_uuid_release=""
-application/code_sign_identity_release=""
-application/export_method_release=0
+application/code_sign_identity_release="iPhone Developer"
+application/export_method_release=1
 application/name=""
 application/info="Made with Godot Engine"
 application/identifier="org.abbysoft.fortseige"
@@ -67,3 +67,53 @@ portrait_launch_screens/iphone_1125x2436=""
 portrait_launch_screens/ipad_768x1024=""
 portrait_launch_screens/ipad_1536x2048=""
 portrait_launch_screens/iphone_1242x2208=""
+
+[preset.1]
+
+name="HTML5"
+platform="HTML5"
+runnable=true
+custom_features=""
+export_filter="all_resources"
+include_filter=""
+exclude_filter=""
+export_path="builds/html/fort_seige.html"
+script_export_mode=1
+script_encryption_key=""
+
+[preset.1.options]
+
+custom_template/debug=""
+custom_template/release=""
+variant/export_type=0
+vram_texture_compression/for_desktop=true
+vram_texture_compression/for_mobile=false
+html/custom_html_shell=""
+html/head_include=""
+html/canvas_resize_policy=2
+html/experimental_virtual_keyboard=false
+
+[preset.2]
+
+name="Linux/X11"
+platform="Linux/X11"
+runnable=true
+custom_features=""
+export_filter="all_resources"
+include_filter=""
+exclude_filter=""
+export_path="builds/linux/fort_seige.x86_64"
+script_export_mode=1
+script_encryption_key=""
+
+[preset.2.options]
+
+custom_template/debug=""
+custom_template/release=""
+binary_format/64_bits=true
+binary_format/embed_pck=false
+texture_format/bptc=false
+texture_format/s3tc=true
+texture_format/etc=false
+texture_format/etc2=false
+texture_format/no_bptc_fallbacks=true
