+++ scenes/game.tscn
@@ -1,0 +1,233 @@
+[gd_scene load_steps=25 format=2]
+
+[ext_resource path="res://graphics/TrajectoryPainter.tscn" type="PackedScene" id=1]
+[ext_resource path="res://Background.tscn" type="PackedScene" id=2]
+[ext_resource path="res://Catapult.tscn" type="PackedScene" id=3]
+[ext_resource path="res://cannon.tscn" type="PackedScene" id=4]
+[ext_resource path="res://EnemyCounterPanel.tscn" type="PackedScene" id=5]
+[ext_resource path="res://TimerWidget.tscn" type="PackedScene" id=6]
+[ext_resource path="res://res/Full-Background.png" type="Texture" id=7]
+[ext_resource path="res://menu.tscn" type="PackedScene" id=8]
+[ext_resource path="res://CoinCounter.tscn" type="PackedScene" id=9]
+[ext_resource path="res://widgets/LevelSlider.tscn" type="PackedScene" id=10]
+[ext_resource path="res://CoinSpawner.tscn" type="PackedScene" id=11]
+[ext_resource path="res://res/battle.mp3" type="AudioStream" id=12]
+[ext_resource path="res://res/wood_wall.png" type="Texture" id=13]
+[ext_resource path="res://addons/PerfMonitor/MonitorPlotPanelPerf.tscn" type="PackedScene" id=14]
+[ext_resource path="res://ShotCounter.tscn" type="PackedScene" id=15]
+[ext_resource path="res://StaticToCamera.cs" type="Script" id=16]
+[ext_resource path="res://res/Cinzel/Cinzel-VariableFont_wght.ttf" type="DynamicFontData" id=17]
+[ext_resource path="res://PropertyBasedActivator.cs" type="Script" id=18]
+[ext_resource path="res://res/Cinzel/static/Cinzel-Bold.ttf" type="DynamicFontData" id=19]
+[ext_resource path="res://scenes/Game.cs" type="Script" id=20]
+[ext_resource path="res://LevelLabel.cs" type="Script" id=24]
+
+[sub_resource type="RectangleShape2D" id=1]
+extents = Vector2( 1492.51, 47.8794 )
+
+[sub_resource type="DynamicFont" id=2]
+size = 65
+font_data = ExtResource( 19 )
+
+[sub_resource type="DynamicFont" id=3]
+size = 75
+outline_size = 5
+outline_color = Color( 0, 0, 0, 1 )
+use_filter = true
+font_data = ExtResource( 17 )
+
+[node name="Game" type="Node2D"]
+script = ExtResource( 20 )
+
+[node name="TextureRect" type="TextureRect" parent="."]
+margin_left = -2.0
+margin_top = 174.0
+margin_right = 2430.0
+margin_bottom = 1393.0
+mouse_filter = 2
+texture = ExtResource( 7 )
+expand = true
+stretch_mode = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="Backgound" parent="." instance=ExtResource( 2 )]
+position = Vector2( -0.785157, -0.523132 )
+
+[node name="BuildingPosition" type="Position2D" parent="."]
+position = Vector2( 1537.01, 1241.02 )
+
+[node name="Weapons" type="Node2D" parent="."]
+position = Vector2( 188.253, 1208.75 )
+script = ExtResource( 18 )
+property = "CurrentWeapon"
+
+[node name="Catapult" parent="Weapons" instance=ExtResource( 3 )]
+visible = false
+position = Vector2( 0, -30 )
+
+[node name="Cannon" parent="Weapons" instance=ExtResource( 4 )]
+visible = false
+position = Vector2( -64, -50 )
+testVelocity = Vector2( 0, 0 )
+
+[node name="CoinSpawner" parent="." instance=ExtResource( 11 )]
+position = Vector2( 56, 15 )
+
+[node name="StaticBody2D" type="StaticBody2D" parent="."]
+position = Vector2( 1, 1241 )
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="StaticBody2D"]
+position = Vector2( 1209.85, 51.9258 )
+shape = SubResource( 1 )
+
+[node name="TrajectoryDisplayer" parent="." instance=ExtResource( 1 )]
+color = Color( 1, 1, 1, 0 )
+color = Color( 1, 1, 1, 0 )
+
+[node name="UILayout" type="Node2D" parent="."]
+script = ExtResource( 16 )
+
+[node name="FireButton" type="Node2D" parent="UILayout"]
+position = Vector2( 77.54, 150 )
+
+[node name="TouchScreenButton" type="TouchScreenButton" parent="UILayout/FireButton"]
+position = Vector2( -25.243, 93.9332 )
+rotation = -1.57079
+scale = Vector2( 1.87651, 1.84741 )
+normal = ExtResource( 13 )
+
+[node name="Label" type="Label" parent="UILayout/FireButton"]
+margin_left = 1.0
+margin_right = 249.0
+margin_bottom = 74.0
+custom_fonts/font = SubResource( 2 )
+custom_colors/font_color_shadow = Color( 0, 0, 0, 1 )
+text = "Set"
+align = 1
+valign = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="ShotCounter" parent="UILayout" instance=ExtResource( 15 )]
+position = Vector2( 659.517, 37.9242 )
+
+[node name="EnemyCounterPanel" parent="UILayout" instance=ExtResource( 5 )]
+position = Vector2( 1132.14, 10.1646 )
+scale = Vector2( 3, 3 )
+
+[node name="TimerWidget" parent="UILayout" instance=ExtResource( 6 )]
+position = Vector2( 1804.38, 11.7001 )
+scale = Vector2( 2.4, 2.4 )
+
+[node name="LevelLabel" type="Label" parent="UILayout"]
+margin_left = 50.8669
+margin_top = 29.6916
+margin_right = 498.867
+margin_bottom = 131.692
+rect_pivot_offset = Vector2( 51.4364, 74.2216 )
+custom_fonts/font = SubResource( 3 )
+text = "Level name"
+valign = 1
+script = ExtResource( 24 )
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="CoinCounter" parent="UILayout" instance=ExtResource( 9 )]
+position = Vector2( 2187.24, 82.5874 )
+scale = Vector2( 3, 3 )
+
+[node name="LevelSlider" parent="UILayout" instance=ExtResource( 10 )]
+visible = false
+margin_left = 131.664
+margin_top = 321.214
+margin_right = 257.664
+margin_bottom = 881.213
+color = Color( 1, 1, 1, 0.384314 )
+
+[node name="MonitorPlotPanelPerf" parent="UILayout" instance=ExtResource( 14 )]
+margin_left = 1994.95
+margin_top = -0.700806
+margin_right = 2450.95
+margin_bottom = 282.299
+
+[node name="Menu" parent="UILayout" instance=ExtResource( 8 )]
+visible = false
+
+[node name="LevelDone" type="Label" parent="UILayout"]
+visible = false
+margin_top = 283.0
+margin_right = 2456.0
+margin_bottom = 771.0
+custom_fonts/font = SubResource( 3 )
+custom_colors/font_color = Color( 1, 0, 0, 1 )
+text = "Victory"
+align = 1
+valign = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="RestartButton" type="Node2D" parent="UILayout"]
+visible = false
+position = Vector2( 1129.07, 622.549 )
+
+[node name="HBoxContainer" type="HBoxContainer" parent="UILayout/RestartButton"]
+margin_left = -1127.0
+margin_top = -27.0
+margin_right = 1329.0
+margin_bottom = 133.0
+size_flags_horizontal = 3
+size_flags_vertical = 3
+alignment = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="MarginContainer" type="MarginContainer" parent="UILayout/RestartButton/HBoxContainer"]
+margin_left = 1053.0
+margin_right = 1403.0
+margin_bottom = 160.0
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="MarginContainer" type="MarginContainer" parent="UILayout/RestartButton/HBoxContainer/MarginContainer"]
+margin_right = 350.0
+margin_bottom = 160.0
+custom_constants/margin_right = 200
+custom_constants/margin_top = 0
+custom_constants/margin_left = 150
+
+[node name="TouchScreenButton" type="TouchScreenButton" parent="UILayout/RestartButton/HBoxContainer/MarginContainer/MarginContainer"]
+position = Vector2( -29.0125, 131.326 )
+rotation = -1.57079
+scale = Vector2( 1.97527, 2.45541 )
+normal = ExtResource( 13 )
+
+[node name="Label" type="Label" parent="UILayout/RestartButton/HBoxContainer/MarginContainer"]
+margin_top = 35.0
+margin_right = 350.0
+margin_bottom = 124.0
+custom_fonts/font = SubResource( 2 )
+custom_colors/font_color_shadow = Color( 0, 0, 0, 1 )
+text = "Restart"
+align = 1
+valign = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="BackgroundMusic" type="AudioStreamPlayer2D" parent="."]
+stream = ExtResource( 12 )
+volume_db = -6.674
+
+[connection signal="LevelChanged" from="." to="UILayout/LevelLabel" method="LevelChanged"]
+[connection signal="pressed" from="UILayout/FireButton/TouchScreenButton" to="." method="FireButtonPressed"]
+[connection signal="NoMoreShots" from="UILayout/ShotCounter" to="." method="NoMoreShots"]
+[connection signal="Timeout" from="UILayout/TimerWidget" to="." method="onTimeout"]
+[connection signal="StartGame" from="UILayout/Menu" to="." method="OnMenuGameStartedSignal"]
+[connection signal="pressed" from="UILayout/RestartButton/HBoxContainer/MarginContainer/MarginContainer/TouchScreenButton" to="." method="OnLevelOverButtonPressed"]
