+++ textures/peasant_preview.tres
@@ -1,0 +1,11 @@
+[gd_resource type="AtlasTexture" load_steps=3 format=2]
+
+[ext_resource path="res://res/peasant_preview.svg" type="Texture" id=1]
+
+[sub_resource type="AtlasTexture" id=1]
+flags = 4
+atlas = ExtResource( 1 )
+
+[resource]
+flags = 4
+atlas = SubResource( 1 )
