+++ ui/Map.tscn
@@ -1,6 +1,6 @@
-[gd_scene load_steps=12 format=2]
+[gd_scene load_steps=14 format=2]
 
-[ext_resource path="res://res/map.svg" type="Texture" id=1]
+[ext_resource path="res://res/map.png" type="Texture" id=1]
 [ext_resource path="res://res/stone.svg" type="Texture" id=2]
 [ext_resource path="res://res/catapult.svg" type="Texture" id=3]
 [ext_resource path="res://res/Cinzel/Cinzel-VariableFont_wght.ttf" type="DynamicFontData" id=4]
@@ -9,6 +9,7 @@
 [ext_resource path="res://PropertyBasedActivator.cs" type="Script" id=7]
 [ext_resource path="res://res/knight.svg" type="Texture" id=8]
 [ext_resource path="res://ui/GlobalMap.cs" type="Script" id=9]
+[ext_resource path="res://res/peasant_preview.svg" type="Texture" id=10]
 
 [sub_resource type="DynamicFont" id=1]
 size = 64
@@ -23,14 +24,19 @@ corner_radius_bottom_right = 50
 corner_radius_bottom_left = 50
 corner_detail = 9
 
+[sub_resource type="AtlasTexture" id=3]
+flags = 4
+atlas = ExtResource( 10 )
+region = Rect2( 1054.42, 138.987, 1631.31, 2259.05 )
+
 [node name="LevelMap" type="Node2D"]
 script = ExtResource( 9 )
 
 [node name="TextureRect" type="TextureRect" parent="."]
-margin_left = -106.12
-margin_top = -38.1384
-margin_right = 2597.88
-margin_bottom = 1352.86
+margin_left = -84.6853
+margin_top = 25.082
+margin_right = 2708.32
+margin_bottom = 1540.08
 texture = ExtResource( 1 )
 expand = true
 __meta__ = {
@@ -38,9 +44,10 @@ __meta__ = {
 }
 
 [node name="Levels" type="Node2D" parent="."]
+position = Vector2( 34, 142 )
 
 [node name="Level1" type="Node2D" parent="Levels"]
-position = Vector2( 274.478, 996.596 )
+position = Vector2( 273.478, 1003.6 )
 
 [node name="Label" type="Label" parent="Levels/Level1"]
 margin_left = 67.2226
@@ -55,28 +62,28 @@ __meta__ = {
 }
 
 [node name="PassedMark4" type="Panel" parent="Levels/Level1"]
-margin_left = -35.669
-margin_top = -37.253
-margin_right = 36.331
-margin_bottom = 35.744
+margin_left = -38.0
+margin_top = -39.809
+margin_right = 42.0
+margin_bottom = 41.191
 custom_styles/panel = SubResource( 2 )
 __meta__ = {
 "_edit_use_anchors_": false
 }
 
 [node name="TextureRect" type="TextureRect" parent="Levels/Level1"]
-margin_left = -86.7085
-margin_top = -229.909
-margin_right = 58.2915
-margin_bottom = 15.0914
-texture = ExtResource( 8 )
+margin_left = -93.0
+margin_top = -230.0
+margin_right = 85.0
+margin_bottom = 15.0
+texture = SubResource( 3 )
 expand = true
 __meta__ = {
 "_edit_use_anchors_": false
 }
 
 [node name="Level2" type="Node2D" parent="Levels"]
-position = Vector2( 502.842, 830.111 )
+position = Vector2( 508.842, 825.111 )
 
 [node name="Label2" type="Label" parent="Levels/Level2"]
 margin_left = -306.282
@@ -91,29 +98,29 @@ __meta__ = {
 "_edit_use_anchors_": false
 }
 
-[node name="PassedMark3" type="Panel" parent="Levels/Level2"]
-margin_left = -33.1396
-margin_top = -37.1587
-margin_right = 38.8604
-margin_bottom = 35.8383
+[node name="PassedMark4" type="Panel" parent="Levels/Level2"]
+margin_left = -38.0
+margin_top = -39.809
+margin_right = 42.0
+margin_bottom = 41.191
 custom_styles/panel = SubResource( 2 )
 __meta__ = {
 "_edit_use_anchors_": false
 }
 
-[node name="TextureRect" type="TextureRect" parent="Levels/Level2"]
-margin_left = -86.7085
-margin_top = -229.909
-margin_right = 58.2915
-margin_bottom = 15.0914
-texture = ExtResource( 8 )
+[node name="TextureRect2" type="TextureRect" parent="Levels/Level2"]
+margin_left = -93.0
+margin_top = -230.0
+margin_right = 85.0
+margin_bottom = 15.0
+texture = SubResource( 3 )
 expand = true
 __meta__ = {
 "_edit_use_anchors_": false
 }
 
 [node name="Level3" type="Node2D" parent="Levels"]
-position = Vector2( 716.947, 638.58 )
+position = Vector2( 729.947, 612.58 )
 
 [node name="Label3" type="Label" parent="Levels/Level3"]
 margin_left = 41.7699
@@ -127,11 +134,11 @@ __meta__ = {
 "_edit_use_anchors_": false
 }
 
-[node name="PassedMark2" type="Panel" parent="Levels/Level3"]
-margin_left = -34.3901
-margin_top = -33.166
-margin_right = 37.6099
-margin_bottom = 39.8309
+[node name="PassedMark4" type="Panel" parent="Levels/Level3"]
+margin_left = -39.0
+margin_top = -36.809
+margin_right = 41.0
+margin_bottom = 44.191
 custom_styles/panel = SubResource( 2 )
 __meta__ = {
 "_edit_use_anchors_": false
@@ -149,7 +156,7 @@ __meta__ = {
 }
 
 [node name="Level4" type="Node2D" parent="Levels"]
-position = Vector2( 961.044, 768.232 )
+position = Vector2( 981.044, 760.232 )
 
 [node name="Label4" type="Label" parent="Levels/Level4"]
 margin_left = 47.2155
@@ -163,11 +170,11 @@ __meta__ = {
 "_edit_use_anchors_": false
 }
 
-[node name="PassedMark" type="Panel" parent="Levels/Level4"]
-margin_left = -36.0396
-margin_top = -39.921
-margin_right = 35.9604
-margin_bottom = 33.076
+[node name="PassedMark4" type="Panel" parent="Levels/Level4"]
+margin_left = -38.0
+margin_top = -39.809
+margin_right = 42.0
+margin_bottom = 41.191
 custom_styles/panel = SubResource( 2 )
 __meta__ = {
 "_edit_use_anchors_": false
@@ -185,18 +192,19 @@ __meta__ = {
 }
 
 [node name="WeaponIcons" type="Node2D" parent="."]
-position = Vector2( 117, 989 )
+position = Vector2( 151, 1131 )
 script = ExtResource( 7 )
 property = "CurrentWeapon"
 
 [node name="Catapult" type="Node2D" parent="WeaponIcons"]
-position = Vector2( -28, 2 )
+position = Vector2( -6, 4 )
 
 [node name="Ball" type="TextureRect" parent="WeaponIcons/Catapult"]
 margin_left = -83.2228
 margin_top = -81.7336
 margin_right = -33.2228
 margin_bottom = -31.7336
+rect_scale = Vector2( 1.00332, 1 )
 texture = ExtResource( 2 )
 expand = true
 __meta__ = {
@@ -218,7 +226,7 @@ __meta__ = {
 [node name="Catapult" type="TextureRect" parent="WeaponIcons/Catapult"]
 margin_left = -108.789
 margin_top = -81.4631
-margin_right = 147.211
+margin_right = 91.211
 margin_bottom = 66.5369
 texture = ExtResource( 3 )
 expand = true
@@ -226,8 +234,10 @@ __meta__ = {
 "_edit_use_anchors_": false
 }
 
-[node name="Cannon" type="TextureRect" parent="WeaponIcons"]
+[node name="Cannon" type="Node2D" parent="WeaponIcons"]
 visible = false
+
+[node name="Cannon" type="TextureRect" parent="WeaponIcons/Cannon"]
 margin_left = -129.758
 margin_top = -50.8904
 margin_right = 132.242
